consolidate data structures
