+++ src/bot/solve.py
@@ -1,14 +1,14 @@
 """Backtest the trading strategy."""
 
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 import subprocess
-from typing import Any, Generator
+from typing import Any
 import numpy as np
 import pandas as pd
 import talib
 import v20  # type: ignore
 from numba import jit  # type: ignore
+from bot.common import BacktestResult, ChartConfig, SolverConfig
 from core.chart import heiken_ashi_numpy
 from numpy.typing import NDArray
 
@@ -19,8 +19,6 @@ from bot.exchange import (
 )
 
 import logging
-
-APP_START_TIME = datetime.now()
 
 
 def get_git_info() -> tuple[str, bool] | Exception:
@@ -46,107 +44,6 @@ def get_git_info() -> tuple[str, bool] | Exception:
         return commit_hash, porcelain_status == ""
     except subprocess.CalledProcessError as e:
         return e
-
-
-class PerfTimer:
-    """PerfTimer class."""
-
-    def __init__(self, app_start_time: datetime, logger: logging.Logger):
-        """Initialize a PerfTimer object."""
-        self.app_start_time = app_start_time
-        self.logger = logger
-        pass
-
-    def __enter__(self):
-        """Start the timer."""
-        self.start = datetime.now()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        """Stop the timer."""
-        self.end = datetime.now()
-        self.logger.info(f"run interval: {self.end - self.start}")
-        self.logger.info("up time: %s", (self.end - self.app_start_time))
-        self.logger.info("last run time: %s", self.end.strftime("%Y-%m-%d %H:%M:%S"))
-
-
-@dataclass
-class ChartConfig:
-    """ChartConfig class."""
-
-    instrument: str
-    granularity: str
-    candle_count: int
-    datefrom: datetime | None = None
-
-
-@dataclass
-class SolverConfig:
-    """SolverConfig class."""
-
-    take_profit: list[float]
-    stop_loss: list[float]
-    source_columns: list[str]
-
-    def get_configs(
-        self, kernel_conf: KernelConfig
-    ) -> tuple[Generator[KernelConfig], int]:
-        """Get column pairs."""
-        if kernel_conf.signal_buy_column == "":
-            gen = (
-                KernelConfig(
-                    signal_buy_column=sb,
-                    signal_exit_column=se,
-                    source_column=so,
-                    take_profit=tp,
-                    stop_loss=sl,
-                    wma_period=kernel_conf.wma_period,
-                )
-                for so in self.source_columns
-                for sb in self.source_columns
-                for se in self.source_columns
-                for tp in self.take_profit
-                for sl in self.stop_loss
-            )
-            return gen, len(self.source_columns) ** 3 * len(self.take_profit) * len(
-                self.stop_loss
-            )
-        else:
-            gen = (
-                KernelConfig(
-                    signal_buy_column=kernel_conf.signal_buy_column,
-                    signal_exit_column=kernel_conf.signal_exit_column,
-                    source_column=kernel_conf.source_column,
-                    wma_period=kernel_conf.wma_period,
-                    take_profit=tp,
-                    stop_loss=sl,
-                )
-                for tp in self.take_profit
-                for sl in self.stop_loss
-            )
-            return gen, len(self.take_profit) * len(self.stop_loss)
-
-
-@dataclass
-class BacktestResult:
-    """BacktestResult class."""
-
-    instrument: str
-    kernel_conf: KernelConfig
-    exit_total: np.float64
-    ratio: np.float64
-    wins: np.int64
-    losses: np.int64
-
-    def __str__(self):
-        """Return a string representation of the BacktestResult object."""
-        return (
-            f"result: {self.kernel_conf} "
-            f"et:{round(self.exit_total, 5)} "
-            f"r:{round(self.ratio, 5)} "
-            f"wins:{self.wins} "
-            f"losses:{self.losses}"
-        )
 
 
 def preprocess(
